Update error message for overread packets

+++ src/netbase.cpp
@@ -532,8 +532,8 @@ int netbase::fireCallbacks( vector<netpacket*>& packets) {
                         << conBufferIndex[con] << "/" << conBufferLength[con] << endl;
                     cerr << "ERROR!: Read past end of packet." << endl;
                     cerr << "bytes_read=" << bytes_read << " Index was "
-                        << (conBufferIndex[con] - bytes_read) << " first byte=0x"
-                        << hex << conBuffer[ (conBufferIndex[con] - bytes_read) ]
+                        << (int)(conBufferIndex[con] - bytes_read) << " first byte=0x"
+                        << hex << (int)conBuffer[ (conBufferIndex[con] - bytes_read) ]
                         << dec << endl;
                     
                     //Set index back to max length and quit
